requirements.txt
anthropic>=0.39.0
streamlit>=1.39.0
python-dotenv>=1.0.0
sqlalchemy>=2.0.0
pandas>=2.0.0
plotly>=5.18.0
pydantic>=2.0.0
